Refactor meta processors code structure
