+++ src/meta/CMakeLists.txt
@@ -12,31 +12,31 @@ add_library(
     MetaServiceHandler.cpp
     MetaServiceUtils.cpp
     MetaHttpHandler.cpp
-    processors/AddHostsProcessor.cpp
-    processors/ListHostsProcessor.cpp
-    processors/RemoveHostsProcessor.cpp
-    processors/CreateSpaceProcessor.cpp
-    processors/GetSpaceProcessor.cpp
-    processors/ListSpacesProcessor.cpp
-    processors/DropSpaceProcessor.cpp
-    processors/GetPartsAllocProcessor.cpp
-    processors/CreateTagProcessor.cpp
-    processors/AlterTagProcessor.cpp
-    processors/GetTagProcessor.cpp
-    processors/ListTagsProcessor.cpp
-    processors/DropTagProcessor.cpp
-    processors/CreateEdgeProcessor.cpp
-    processors/AlterEdgeProcessor.cpp
-    processors/GetEdgeProcessor.cpp
-    processors/ListEdgesProcessor.cpp
-    processors/DropEdgeProcessor.cpp
-    processors/GetProcessor.cpp
-    processors/MultiGetProcessor.cpp
-    processors/MultiPutProcessor.cpp
-    processors/RemoveProcessor.cpp
-    processors/RemoveRangeProcessor.cpp
-    processors/ScanProcessor.cpp
-    processors/HBProcessor.cpp
+    processors/partsMan/AddHostsProcessor.cpp
+    processors/partsMan/ListHostsProcessor.cpp
+    processors/partsMan/RemoveHostsProcessor.cpp
+    processors/partsMan/CreateSpaceProcessor.cpp
+    processors/partsMan/GetSpaceProcessor.cpp
+    processors/partsMan/ListSpacesProcessor.cpp
+    processors/partsMan/DropSpaceProcessor.cpp
+    processors/partsMan/GetPartsAllocProcessor.cpp
+    processors/schemaMan/CreateTagProcessor.cpp
+    processors/schemaMan/AlterTagProcessor.cpp
+    processors/schemaMan/GetTagProcessor.cpp
+    processors/schemaMan/ListTagsProcessor.cpp
+    processors/schemaMan/DropTagProcessor.cpp
+    processors/schemaMan/CreateEdgeProcessor.cpp
+    processors/schemaMan/AlterEdgeProcessor.cpp
+    processors/schemaMan/GetEdgeProcessor.cpp
+    processors/schemaMan/ListEdgesProcessor.cpp
+    processors/schemaMan/DropEdgeProcessor.cpp
+    processors/customKV/GetProcessor.cpp
+    processors/customKV/MultiGetProcessor.cpp
+    processors/customKV/MultiPutProcessor.cpp
+    processors/customKV/RemoveProcessor.cpp
+    processors/customKV/RemoveRangeProcessor.cpp
+    processors/customKV/ScanProcessor.cpp
+    processors/admin/HBProcessor.cpp
 )
 add_dependencies(
     meta_service_handler
