+++ src/meta/processors/customKV/GetProcessor.cpp
@@ -5,7 +5,7 @@
  */
 
 
-#include "meta/processors/GetProcessor.h"
+#include "meta/processors/customKV/GetProcessor.h"
 
 namespace nebula {
 namespace meta {
