+++ src/meta/processors/partsMan/AddHostsProcessor.cpp
@@ -4,20 +4,20 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#include "meta/processors/MultiPutProcessor.h"
+#include "meta/processors/partsMan/AddHostsProcessor.h"
 
 namespace nebula {
 namespace meta {
 
-void MultiPutProcessor::process(const cpp2::MultiPutReq& req) {
-    CHECK_SEGMENT(req.get_segment());
+void AddHostsProcessor::process(const cpp2::AddHostsReq& req) {
+    folly::SharedMutex::WriteHolder wHolder(LockUtils::spaceLock());
     std::vector<kvstore::KV> data;
-    for (auto& pair : req.get_pairs()) {
-        data.emplace_back(MetaServiceUtils::assembleSegmentKey(req.get_segment(), pair.get_key()),
-                          pair.get_value());
+    for (auto& h : req.get_hosts()) {
+        data.emplace_back(MetaServiceUtils::hostKey(h.ip, h.port), MetaServiceUtils::hostVal());
     }
     doPut(std::move(data));
 }
 
 }  // namespace meta
 }  // namespace nebula
+
