+++ src/meta/processors/partsMan/GetPartsAllocProcessor.cpp
@@ -4,7 +4,7 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#include "meta/processors/GetPartsAllocProcessor.h"
+#include "meta/processors/partsMan/GetPartsAllocProcessor.h"
 
 namespace nebula {
 namespace meta {
