+++ src/meta/processors/partsMan/GetSpaceProcessor.cpp
@@ -4,7 +4,7 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#include "meta/processors/GetSpaceProcessor.h"
+#include "meta/processors/partsMan/GetSpaceProcessor.h"
 
 namespace nebula {
 namespace meta {
