+++ src/meta/processors/schemaMan/AlterTagProcessor.cpp
@@ -4,7 +4,7 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#include "meta/processors/AlterTagProcessor.h"
+#include "meta/processors/schemaMan/AlterTagProcessor.h"
 #include "time/TimeUtils.h"
 
 namespace nebula {
