+++ src/meta/processors/schemaMan/DropEdgeProcessor.cpp
@@ -4,7 +4,7 @@
  * attached with Common Clause Condition 1.0, found in the LICENSES directory.
  */
 
-#include "meta/processors/DropEdgeProcessor.h"
+#include "meta/processors/schemaMan/DropEdgeProcessor.h"
 
 namespace nebula {
 namespace meta {
