+++ src/meta/test/HBProcessorTest.cpp
@@ -11,7 +11,7 @@
 #include "fs/TempDir.h"
 #include "meta/test/TestUtils.h"
 #include <common/time/TimeUtils.h>
-#include "meta/processors/HBProcessor.h"
+#include "meta/processors/admin/HBProcessor.h"
 
 DECLARE_int32(expired_hosts_check_interval_sec);
 DECLARE_int32(expired_threshold_sec);
